Made slightly-delayed (in the 2-7ms range) versions of the pistol and minigun fire sounds
If you noticed the pistols and minigun sounding muddy in GZDoom and not in Zandronum, it
turns out that FMod seems to introduce a very slight delay in sounds - I'm not sure if
this is by design or on accident, but it makes the pistols sound a lot less muddy and
more natural with their looping. OpenAL is much more exact with when sounds play, and
that's why they sound bad. This explicitly reintroduces that delay, although it does make
the minigun sound awful in Zandronum due to FMod apparently clipping sounds a bit earlier.

I intend to redo the minigun to use a looping sound because of this; there's simply no
dealing with the issue any other way, it seems.