The shotgun now pumps if you have ammo for either fire mode, rather than just the one you're using
Also now it never switches off when out of ammo. I was gonna do something like
Quake 3 empty switching, but doing asynchronous weapon stuff in Zandronum 3 is
like pulling teeth, so I don't really want to bother. And chances are I'm not
gonna be recoding the weapons if/when Zandronum finally does get ZScript (or
even things like A_Overlay), so I guess I just have to live with it.