Nerfed alt ammo amount in bullet/shell boxes, buffed flamer fuel count in cell packs
The alternate ammo types are just as powerful as rockets and cells are, and
that's because they exist to keep the shotgun/SSG/minigun relevant when the
impaler and lancer take center stage. The amount of alternate ammo you get
from bullet and shell boxes allows you to use them in the big fights, but
not the small stuff in between, in the hopes to preserve the map's intended
pacing at least a little bit.

I mean come on, it's already easy enough as is without being able to burn
every ambush to a crisp, or tear them apart with a wave of exploding shrapnel.
I can't make it *too* easy, or else it just gets boring.